backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.44
psycopg[binary]==3.2.13
alembic==1.15.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pydantic-settings==2.8.1
reportlab==4.2.5
